+++ src/python/impactx/dashboard/Input/space_charge/__init__.py
@@ -1,0 +1,5 @@
+from .utils import SpaceChargeFunctions
+
+__all__ = [
+    "SpaceChargeFunctions",
+]
