--- src/python/impactx/dashboard/Input/space_charge_configuration/__init__.py
@@ -1,5 +1,0 @@
-from .spaceChargeFunctions import SpaceChargeFunctions
-
-__all__ = [
-    "SpaceChargeFunctions",
-]
